Send sigterm before sigkill on tomcat stop

+++ kamehouse-shell/bin/lin/kamehouse/tomcat-stop.sh
@@ -9,10 +9,25 @@ mainProcess() {
   TOMCAT_PID=`netstat -nltp | grep ":${TOMCAT_PORT} " | grep java | awk '{print $7}' | cut -d '/' -f 1`
   if [ -z "${TOMCAT_PID}" ]; then
     log.info "Tomcat is not running"
-  else
-    log.info "Killing process ${COL_PURPLE}${TOMCAT_PID}"
-    kill -9 ${TOMCAT_PID}
+    return
   fi
+  log.info "Killing process ${COL_PURPLE}${TOMCAT_PID}"
+  kill ${TOMCAT_PID}
+
+  log.info "Waiting for for tomcat to exit..."
+  MAX_WAIT=20
+  COUNT=0
+  while kill -0 ${TOMCAT_PID} 2>/dev/null && [ $COUNT -lt $MAX_WAIT ]; do
+    sleep 1
+    ((COUNT++))
+  done
+
+  if kill -0 ${TOMCAT_PID} 2>/dev/null; then
+    log.warn "Tomcat did not stop gracefully after ${MAX_WAIT}s. Forcing kill."
+    kill -9 ${TOMCAT_PID}
+  else
+    log.info "Tomcat stopped successfully."
+  fi  
 }
 
 parseArguments() {
